mcp/requirements.txt
flask
python-dotenv
requests
faiss-cpu
pdfminer.six
tqdm
tinydb